Split ConnectionSteps
