+++ test/behaviour/concept/type/thingtype/ThingTypeSteps.java
@@ -31,10 +31,9 @@ import io.cucumber.java.en.When;
 
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Stream;
 
 import static grakn.client.test.behaviour.config.Parameters.RootLabel;
-import static grakn.client.test.behaviour.connection.ConnectionSteps.tx;
+import static grakn.client.test.behaviour.connection.ConnectionStepsCore.tx;
 import static grakn.client.test.behaviour.util.Util.assertThrows;
 import static java.util.Objects.isNull;
 import static java.util.stream.Collectors.toSet;
