+++ test/behaviour/connection/session/BUILD
@@ -29,7 +29,7 @@ java_library(
     deps = [
         # Internal Package Dependencies
         "//:client-java",
-        "//test/behaviour/connection:steps",
+        "//test/behaviour/connection:steps-core",
 
         # Internal Repository Dependencies
         "@graknlabs_common//:common",
