+++ test/behaviour/connection/transaction/BUILD
@@ -29,7 +29,30 @@ java_library(
     deps = [
         # Internal Package Dependencies
         "//:client-java",
-        "//test/behaviour/connection:steps",
+        "//test/behaviour/connection:steps-core",
+        "//test/behaviour/util:util",
+
+        # Internal Repository Dependencies
+        "@graknlabs_common//:common",
+        "@graknlabs_graql//java:graql",
+
+        # External Maven Dependencies
+        "@maven//:junit_junit",
+        "@maven//:io_cucumber_cucumber_java",
+        "@maven//:org_hamcrest_hamcrest_library",
+    ],
+    visibility = ["//visibility:public"],
+)
+
+java_library(
+    name = "steps-cluster",
+    srcs = [
+        "TransactionSteps.java"
+    ],
+    deps = [
+        # Internal Package Dependencies
+        "//:client-java",
+        "//test/behaviour/connection:steps-core",
         "//test/behaviour/util:util",
 
         # Internal Repository Dependencies
