+++ test/behaviour/connection/transaction/TransactionSteps.java
@@ -32,13 +32,13 @@ import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 
-import static grakn.client.test.behaviour.connection.ConnectionSteps.THREAD_POOL_SIZE;
-import static grakn.client.test.behaviour.connection.ConnectionSteps.sessions;
-import static grakn.client.test.behaviour.connection.ConnectionSteps.sessionsParallel;
-import static grakn.client.test.behaviour.connection.ConnectionSteps.sessionsParallelToTransactionsParallel;
-import static grakn.client.test.behaviour.connection.ConnectionSteps.sessionsToTransactions;
-import static grakn.client.test.behaviour.connection.ConnectionSteps.sessionsToTransactionsParallel;
-import static grakn.client.test.behaviour.connection.ConnectionSteps.threadPool;
+import static grakn.client.test.behaviour.connection.ConnectionStepsCore.THREAD_POOL_SIZE;
+import static grakn.client.test.behaviour.connection.ConnectionStepsCore.sessions;
+import static grakn.client.test.behaviour.connection.ConnectionStepsCore.sessionsParallel;
+import static grakn.client.test.behaviour.connection.ConnectionStepsCore.sessionsParallelToTransactionsParallel;
+import static grakn.client.test.behaviour.connection.ConnectionStepsCore.sessionsToTransactions;
+import static grakn.client.test.behaviour.connection.ConnectionStepsCore.sessionsToTransactionsParallel;
+import static grakn.client.test.behaviour.connection.ConnectionStepsCore.threadPool;
 import static grakn.client.test.behaviour.util.Util.assertThrows;
 import static grakn.client.test.behaviour.util.Util.assertThrowsWithMessage;
 import static grakn.common.collection.Collections.list;
