+++ test/behaviour/graql/BUILD
@@ -27,7 +27,7 @@ java_library(
     ],
     deps = [
         # Package dependencies
-        "//test/behaviour/connection:steps",
+        "//test/behaviour/connection:steps-core",
         "//test/behaviour/util",
         "//:client-java",
